Replace hard-coded int/float/double with custom types
avoids mixing of int/float/double with SLONG/FLOAT/DOUBLE etc

+++ src/TeakLibW/Math.cpp
@@ -1,7 +1,7 @@
 #include "StdAfx.h"
 
 void memswap(void *dst, void *src, ULONG size) {
-    for (unsigned int i = size; i > 0; --i) {
+    for (ULONG i = size; i > 0; --i) {
         Swap((static_cast<BYTE *>(dst))[i - 1], (static_cast<BYTE *>(src))[i - 1]);
     }
 }
@@ -11,15 +11,15 @@ SLONG CalcInertiaVelocity(SLONG a, SLONG b) {
         if (a >= b) {
             return 0;
         }
-        return -static_cast<SLONG>(sqrt(static_cast<double>(b - a + 1) / 2.0 + 0.5));
+        return -static_cast<SLONG>(sqrt(static_cast<DOUBLE>(b - a + 1) / 2.0 + 0.5));
     }
 
-    return static_cast<SLONG>(sqrt(static_cast<double>(a - b + 1) / 2.0 + 0.5));
+    return static_cast<SLONG>(sqrt(static_cast<DOUBLE>(a - b + 1) / 2.0 + 0.5));
 }
 
 SLONG Calc1nSum(SLONG a) { return a * (a + 1) / 2; }
 
-double GetFrameRate() {
+DOUBLE GetFrameRate() {
     TeakLibW_Exception(nullptr, 0, ExcNotImplemented);
     return 42.0;
 }
